refactor(ui): extract inspector panels from Inspector
